Implemented start(), stop(), and update() methods to wave
These will be used to make the wave spawn mobs, etc. as well as to
force-stop it and continually check if it needs to spawn more mobs. In
order to do that, we will need to externally call the active wave's
"update()" method repeatedly after its "start()" method is externally
called.